Add tests for NSString categories

+++ CotEditor/Sources/NSString+Normalization.m
@@ -48,7 +48,7 @@
     }
     
     const char *utf8_src = [self UTF8String];
-    unsigned long length = strlen(utf8_src) * 256;
+    int32_t length = (int32_t)strlen(utf8_src) * 256;
     
     UChar *utf16_src = (UChar*)malloc(sizeof(UChar) * length);
     u_strFromUTF8(utf16_src, length, NULL, utf8_src, -1, &error);
